ActivityMonitoringPy/train/requirements.txt
Flask==1.0.2
firebase-admin==2.13.0
google-cloud-firestore==0.29.0
siganalysis==0.4.0
numpy==1.16.3
matplotlib==3.0.3
scipy==1.2.1 
scikit-learn==0.20.3
